🧬 Migrate the data models about settings in section *apis* to module *api_config.apis*.
