+++ pymock_api/server/application/process.py
@@ -3,7 +3,7 @@ from abc import ABC, ABCMeta, abstractmethod
 from pydoc import locate
 from typing import Any, Dict, List, Union, cast
 
-from ...model.api_config import APIParameter, HTTPRequest, HTTPResponse, MockAPI
+from ...model import APIParameter, HTTPRequest, HTTPResponse, MockAPI
 from ...model.enums import ResponseStrategy
 from .request import BaseCurrentRequest
 from .response import BaseResponse
