+++ pymock_api/server/mock.py
@@ -5,8 +5,7 @@ This module provides objects for mocking APIs as a web application with differen
 
 from typing import Any, Optional
 
-from ..model import APIConfig, load_config
-from ..model.api_config import MockAPIs
+from ..model import APIConfig, MockAPIs, load_config
 from .application import BaseAppServer, FlaskServer
